requirements.txt
pandas==2.0.3
seaborn==0.13.2 
matplotlib==3.7.5
notebook==7.3.3
pyyaml>=6.0.2
easydict==1.10
pillow==9.5.0
numpy==1.21.6
TensorBoardX==2.6
tensorboard==2.14.0
setuptools==59.5.0
opencv-python==4.6.0.66
scipy==1.7.3
scikit-learn==1.0.2
scikit-image==0.19.3
einops==0.6.1
h5py==3.8.0
torch==1.10.1+cu111
torchvision==0.11.2+cu111
mmcv-full==1.4.4
mmsegmentation==0.24.0
timm==0.4.12
ftfy==6.1.1
regex==2023.8.8
openmim==0.3.9
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
-f https://download.pytorch.org/whl/cu111/torch_stable.html 
-f https://download.openmmlab.com/mmcv/dist/cu111/torch1.10.0/index.html